review complete

+++ progetto_lassi/db/migrate/20200614170812_add_telefono_to_user.rb
@@ -4,7 +4,7 @@ class AddTelefonoToUser < ActiveRecord::Migration[5.1]
     add_column :users, :segnalato, :integer, default:0
     add_column :users, :nome, :string
     add_column :users, :cognome, :string
-    add_column :users, :rating, :integer, default:nil
+    add_column :users, :rating, :float, default:nil
     add_column :users, :ratingnum, :integer, default:0 
     add_column :users, :sospeso, :boolean, default:false
     add_column :users, :admin, :boolean, default:false
